feat(create-stagewise-plugin): add ExamplePlugin and refactor A11yPlugin structure
- Introduced ExamplePlugin with a toolbar action for demonstration purposes.
- Refactored A11yPlugin to separate its implementation into a dedicated plugin.tsx file, enhancing code organization and clarity.
