+++ packages/create-stagewise-plugin/template-basic/src/index.ts
@@ -1,0 +1,1 @@
+export { ExamplePlugin } from './plugin';
