+++ packages/create-stagewise-plugin/template-basic/src/plugin.tsx
@@ -1,4 +1,5 @@
 'use client';
+
 import type { ToolbarPlugin } from '@stagewise/toolbar';
 import {
   type FunctionalComponent,
